backend/pyproject.toml
[project]
name = "vercel-fastapi-starter"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "bcrypt>=5.0.0",
  "fastapi[standard]>=0.121.0",
  "psycopg2-binary>=2.9.11",
  "pyjwt>=2.11.0",
  "python-dotenv>=1.2.1",
  "sqlmodel>=0.0.33",
  "uvicorn>=0.34.0",
]
